Move moments directory to _moments
