+++ fairlearn/reductions/_moments/__init__.py
@@ -2,6 +2,7 @@
 # Licensed under the MIT License.
 
 from .moment import Moment  # noqa: F401
+from .moment import ClassificationMoment, LossMoment  # noq: F401
 
 from .conditional_selection_rate import ConditionalSelectionRate  # noqa: F401
 from .conditional_selection_rate import DemographicParity, EqualizedOdds  # noqa: F401
@@ -12,11 +13,13 @@ from .bounded_group_loss import SquareLoss, AbsoluteLoss, ZeroOneLoss  # noqa: F
 __all__ = [
     "AbsoluteLoss",
     "Moment",
+    "ClassificationMoment",
     "ConditionalSelectionRate",
     "DemographicParity",
     "EqualizedOdds",
     "ErrorRate",
     "GroupLossMoment",
+    "LossMoment",
     "SquareLoss",
     "ZeroOneLoss"
 ]
